Add widget API outline, scripts need to be called from the main thread. This needs to be implemented using a custom event. Only works when passed via argument yet.

+++ src/widget/tachometer.c
@@ -61,7 +61,7 @@ static void draw_circle(SDL_Renderer* renderer, int cx, int cy, int radius, bool
     }
 }
 
-void widget_tachometer(uint8 pos_x, uint8 pos_y, uint8 size, const uint32 max, uint32 rpm)
+void widget_tachometer(uint32 pos_x, uint32 pos_y, uint32 size, const uint32 max, uint32 rpm)
 {
     SDL_Renderer* renderer = window_get_renderer();
     uint8 r, g, b;
